completed refactoring of transport tests
