+++ lib/signaling/out-of-band/websocket.js
@@ -10,9 +10,9 @@ var winstonWrapper = require('winston-meta-wrapper')
 
 var SIGNALING_TYPE = 'websocket-signaling'
 
-function WebSocketSignaling (wsOpts) {
+function WebSocketSignaling (args) {
   if (!(this instanceof WebSocketSignaling)) {
-    return new WebSocketSignaling(wsOpts)
+    return new WebSocketSignaling(args)
   }
   // logging
   this._log = winstonWrapper(winston)
@@ -20,10 +20,10 @@ function WebSocketSignaling (wsOpts) {
     module: '1tp:signaling:out-of-band:websocket'
   })
   // init
-  this._opts = merge(Object.create(WebSocketSignaling.DEFAULTS), wsOpts)
+  this._args = merge(Object.create(WebSocketSignaling.DEFAULTS), args)
   AbstractSignaling.call(this)
   // done
-  this._log.debug('created websocket signaling connector with args ' + JSON.stringify(this._opts))
+  this._log.debug('created websocket signaling connector with args ' + JSON.stringify(this._args))
 }
 
 WebSocketSignaling.DEFAULTS = {
@@ -55,9 +55,9 @@ WebSocketSignaling.prototype.register = function (callback, requestedRegistratio
     url = requestedRegistrationInfo.url
   }
   // create random uid if undefined
-  uid = uid || this._opts.uid || hat()
+  uid = uid || this._args.uid || hat()
   // use default websocket url if undefined
-  url = url || this._opts.url
+  url = url || this._args.url
   // create new registration info instance to be returned once registration succeeds
   var registrationInfo = {}
   registrationInfo.type = SIGNALING_TYPE
@@ -65,9 +65,9 @@ WebSocketSignaling.prototype.register = function (callback, requestedRegistratio
   registrationInfo.url = url
   // create socket
   var socket = io.connect(url, {
-    'reconnection delay': this._opts.reconnectionDelay,
-    'reopen delay': this._opts.reopenDelay,
-    'force new connection': this._opts.forceNewConnection
+    'reconnection delay': this._args.reconnectionDelay,
+    'reopen delay': this._args.reopenDelay,
+    'force new connection': this._args.forceNewConnection
   })
   socket.on('connect', this.onConnected(callback, registrationInfo, onSuccess, onFailure))
   socket.on('disconnect', this.onDisconnected())
@@ -186,7 +186,7 @@ WebSocketSignaling.prototype.onPing = function () {
 WebSocketSignaling.prototype.onConnected = function (callback, registrationInfo, onSuccess, onFailure) {
   var self = this
   return function () {
-    self._log.debug('connected to ' + self._opts.url)
+    self._log.debug('connected to ' + self._args.url)
     // send registration message over socket
     var signalingMessage = {}
     signalingMessage.username = registrationInfo.uid
@@ -213,7 +213,7 @@ WebSocketSignaling.prototype.onConnected = function (callback, registrationInfo,
 WebSocketSignaling.prototype.onDisconnected = function () {
   var self = this
   return function () {
-    self._log.debug('disconnected from ' + self._opts.url)
+    self._log.debug('disconnected from ' + self._args.url)
   }
 }
