+++ test/net.unit.js
@@ -23,6 +23,22 @@ var turnUser = process.env.TURN_USER
 var turnPwd = process.env.TURN_PASS
 var registrar = process.env.ONETP_REGISTRAR
 
+if (!turnAddr) {
+  throw new Error ('TURN_ADDR undefined -- giving up')
+}
+if (!turnPort) {
+  throw new Error ('TURN_PORT undefined -- giving up')
+}
+if (!turnUser) {
+  throw new Error ('TURN_USER undefined -- giving up')
+}
+if (!turnPwd) {
+  throw new Error ('TURN_PASS undefined -- giving up')
+}
+if (!registrar) {
+  throw new Error ('ONETP_REGISTRAR undefined -- giving up')
+}
+
 describe('net api', function () {
   this.timeout(20000)
